src/main/java/seminar3/Cart.java
package seminar3;

import java.util.ArrayList;
import java.util.List;

public class Cart {
    private List<Product> products;

    public List<Product> getProducts() {
        return products;
    }

    public Cart() {
        products = new ArrayList<>();
    }

    public void addProduct(Product product) {
        products.add(product);
    }

    public void removeProduct(Product product) {
        products.remove(product);
    }

    public double calculateTotalPrice() {
        double totalPrice = 0;
        for (Product product : products) {
            totalPrice += product.getPrice() * product.getQuantity();
        }
        return totalPrice;
    }

// Другие методы для работы с корзиной
// ...
}

